+++ vite.config.ts
@@ -31,7 +31,10 @@ export default defineConfig(({mode}) => {
     resolve: {
       alias: {
         '@': resolve(__dirname, './src'),
+        _scripts: resolve(__dirname, './src/contentScript/'),
+        _bg: resolve(__dirname, './src/background/'),
         '~': resolve(__dirname, './src/components'),
+        '~~popup': resolve(__dirname, './src/components/modules/popup'),
         $: resolve(__dirname, './src/assets/'),
       },
     },