+++ CHANGELOG.md
@@ -1,5 +1,55 @@
 # CHANGELOG
 
+## 0.1.0
+
+### Minor Changes
+
+- e1b8856: add custom `icons` for extension
+- e1b8856: update extension `details` for alpha release
+
+## 0.0.7
+
+### Patch Changes
+
+- 107fa3b: init `Typography` UI component (dynamic)
+- f2247e6: add `isContrasted` param to `colorExtractor` content script
+- 95f2dcd: add `bullshitFonts` handling to `fontsExtractor` content script
+
+## 0.0.6
+
+### Patch Changes
+
+- 2e08f62: finalize `Header` Global component
+- ec173f9: pass `GET_TAB_INFO` data into `Header` Global component
+- d9eb5ec: init `getTabInfo` background worker for parsing `activeTab` data
+
+## 0.0.5
+
+### Patch Changes
+
+- e1b3ade: refactor `fontsExtractor` and `FontsUnit` popup block
+- e1b3ade: enhance `ColorsUnit` module with `tooltip` and `copy` action
+
+## 0.0.4
+
+### Patch Changes
+
+- 44e4d2f: optimize `Popup` state handling of `Units`
+- e8a0354: init `colorsExtractor` and `ColorsUnit` for its data
+- bb57dc5: integrate `fonts` data into `FontsUnit` popup block
+- 914d80b: init `fontsExtractor` for parsing font data
+- 2f1d6c8: init `FontsUnit` popup block
+- 64f6b7f: init `Unit` UI component
+
+## 0.0.3
+
+### Patch Changes
+
+- 4205568: init `Header` Global component
+- f07b83c: init `Container` Global component
+- 50cf64c: init `Layout` Global component
+- 11be4c6: init `Button` UI component
+
 ## 0.0.2
 
 ### Patch Changes
@@ -12,16 +62,6 @@
 
 - Project configuration completed: CI/CD, Tailwind, Vite & TS setup
 
-```txt
-Summary
-  1. document grouping follow 'SemVer2.0' protocol
-  2. use 'PATCH' as a minimum granularity
-  3. use concise descriptions
-  4. type: feat \ fix \ update \ perf \ remove \ docs \ chore
-  5. version timestamp follow the yyyy.MM.dd format
-```
-
 ## 0.0.0 [2025.01.29]
 
-- feat: initial
-- feat: generator by ![create-chrome-ext](https://github.com/guocaoyi/create-chrome-ext)
+- init `extension` project using `create-chrome-ext`